+++ packages/react/src/PlayProvider.tsx
@@ -1,4 +1,4 @@
-import { PropsWithChildren, createContext, useEffect } from "react";
+import { PropsWithChildren, createContext, useEffect, useState } from "react";
 import playhtml from "./playhtml-singleton";
 import { InitOptions } from "playhtml";
 import { useLocation } from "./hooks/useLocation";
@@ -31,12 +31,23 @@ export function PlayProvider({
 }: PropsWithChildren<Props>) {
   const location = useLocation();
   useEffect(() => {
-    playhtml.init(initOptions);
-  }, []);
-  useEffect(() => {
     // in future migrate this to a more "reactful" way by having all the elements rely state on this context
     playhtml.setupPlayElements();
   }, [location]);
+
+  const [hasSynced, setHasSynced] = useState(false);
+
+  useEffect(() => {
+    playhtml.init(initOptions).then(
+      () => {
+        setHasSynced(true);
+      },
+      (err) => {
+        console.error(err);
+        setHasSynced(true);
+      }
+    );
+  }, []);
 
   return (
     <PlayContext.Provider
@@ -47,7 +58,15 @@ export function PlayProvider({
         removePlayEventListener: playhtml.removePlayEventListener,
       }}
     >
-      {children}
+      {/* <InitPlayhtmlEffect {...props} /> */}
+      {hasSynced ? children : null}
     </PlayContext.Provider>
   );
 }
+
+// function InitPlayhtmlEffect({ initOptions }: Props) {
+//   useEffect(() => {
+//     void playhtml.init(initOptions);
+//   }, []);
+//   return null;
+// }