+++ packages/react/src/index.tsx
@@ -10,9 +10,14 @@ import type {
   ReactElementEventHandlerData,
 } from "./utils";
 
+export type WithPlayProps<T, V> =
+  | Omit<ReactElementInitializer<T, V>, "children">
+  | (Omit<ReactElementInitializer<T, V>, "children" | "defaultData"> & {
+      tagInfo?: Partial<{ [k in TagType]: string }> | TagType[];
+    });
+
 // TODO: make the mapping to for TagType -> ReactElementInitializer
 // TODO: semantically, it should not be `can-play` for all of the pre-defined ones..
-// @deprecated use `withPlay` instead
 export function CanPlayElement<T, V>({
   children,
   id,
@@ -64,6 +69,7 @@ export function CanPlayElement<T, V>({
       // @ts-ignore
       ref.current.updateElementAwareness = updateElement;
       playhtml.setupPlayElement(ref.current, { ignoreIfAlreadySetup: true });
+      // console.log("setting up", ref.current.id);
       const existingData = playhtml.globalData
         ?.get("can-play")
         ?.get(ref.current.id);
@@ -128,6 +134,167 @@ export function CanPlayElement<T, V>({
   );
 }
 
+export type CanPlayProps<
+  T extends object,
+  V = any
+> = ReactElementEventHandlerData<T, V>;
+
+/**
+ * Wrapper to create a higher order component that passes down shared, global state to a component.
+ * You can either pass in a named function or pass an inline one. See examples below:
+ *
+ * interface ColorChange {}
+ * interface Props {}
+ * 1) NAMED FUNCTION
+ * function color(
+ *   props: Props,
+ *   { data, setData }: CanPlayProps<{ colors: ColorChange[] }>
+ * ) {
+ *   return <div>...</div>;
+ * }
+ * export const Color = withSharedState(
+ *   { defaultData: { colors: [] }, },
+ *   color
+ * );
+ *
+ * 2) INLINE FUNCTION
+ * export const ColorInline = withSharedState(
+ *   { defaultData: { colors: [] as ColorChange[] }, },
+ *   ({ data, setData }, props: Props) => {
+ *      return <div>...</div>;
+ *   }
+ * );
+ */
+export function withSharedState<T extends object, P = any, V = any>(
+  playConfig: WithPlayProps<T, V> | ((props: P) => WithPlayProps<T, V>),
+  component: (
+    playProps: ReactElementEventHandlerData<T, V>,
+    props: P
+  ) => React.ReactElement
+): (props: P) => React.ReactElement {
+  const renderChildren = (props: P) => {
+    return (
+      <CanPlayElement
+        tagInfo={undefined}
+        defaultData={undefined}
+        {...(typeof playConfig === "function" ? playConfig(props) : playConfig)}
+      >
+        {(playData) => component(playData, props)}
+      </CanPlayElement>
+    );
+  };
+
+  return (props) => renderChildren(props);
+}
+
+// export function useSharedState<T extends object, V = any>({
+//   id,
+//   ...restProps
+// }: WithPlayProps<T, V> & { id: string }): ReactElementEventHandlerData<T, V> {
+//   const { tagInfo = { "can-play": "" }, ...elementProps } = {
+//     defaultData: undefined,
+//     ...restProps,
+//   };
+//   const computedTagInfo = tagInfo
+//     ? Array.isArray(tagInfo)
+//       ? Object.fromEntries(tagInfo.map((t) => [t, ""]))
+//       : tagInfo
+//     : { "can-play": "" };
+//   const ref = useRef<HTMLElement>(null);
+//   const { defaultData, myDefaultAwareness } = elementProps;
+//   const [data, setData] = useState<T | undefined>(defaultData);
+//   const [awareness, setAwareness] = useState<V[]>(
+//     myDefaultAwareness ? [myDefaultAwareness] : []
+//   );
+//   const [myAwareness, setMyAwareness] = useState<V | undefined>(
+//     myDefaultAwareness
+//   );
+//   // TODO: maybe have a separate one for free-form variables?
+//   playhtml.globalData?.get("can-play")?.set(id, data);
+
+//   // TODO: this is kinda a hack but it works for now since it is called whenever we set data.
+//   const updateElement: ElementInitializer["updateElementAwareness"] = ({
+//     data: newData,
+//     awareness: newAwareness,
+//     myAwareness,
+//   }) => {
+//     setData(newData);
+//     setAwareness(newAwareness);
+//     setMyAwareness(myAwareness);
+//   };
+
+//   useEffect(() => {
+//     if (!ref.current) {
+//       let ele = document.getElementById(id);
+//       if (!ele) {
+//         ele = document.createElement("div");
+//         for (const [tag, value] of Object.entries(computedTagInfo)) {
+//           ele.setAttribute(tag, value);
+//         }
+//         document.body.appendChild(ele).id = id;
+//       }
+//       ref.current = ele;
+//     }
+
+//     for (const [key, value] of Object.entries(elementProps)) {
+//       // @ts-ignore
+//       ref.current[key] = value;
+//     }
+//     // @ts-ignore
+//     ref.current.updateElement = updateElement;
+//     // @ts-ignore
+//     ref.current.updateElementAwareness = updateElement;
+//     playhtml.setupPlayElement(ref.current, { ignoreIfAlreadySetup: true });
+//     console.log("setting up", ref.current.id);
+//     const existingData = playhtml.globalData
+//       ?.get("can-play")
+//       ?.get(ref.current.id);
+//     if (existingData) {
+//       setData(existingData);
+//     }
+//     // console.log("setting up", elementProps.defaultData, ref.current);
+
+//     return () => {
+//       if (!ref.current || !playhtml.elementHandlers) return;
+//       playhtml.removePlayElement(ref.current);
+//     };
+//   }, [playConfig, ref.current]);
+
+//   return {
+//     // @ts-ignore
+//     data,
+//     awareness,
+//     setData: (newData) => {
+//       // console.log("settingdata", newData);
+//       // console.log(ref.current?.id);
+//       // console.log(
+//       //   getCurrentElementHandler(TagType.CanPlay, ref.current?.id || "")
+//       // );
+//       if (!ref.current?.id) {
+//         console.warn(`[@playhtml/react] No id set for element ${ref.current}`);
+//         return;
+//       }
+//       const handler = getCurrentElementHandler(TagType.CanPlay, ref.current.id);
+//       if (!handler) {
+//         console.warn(
+//           `[@playhtml/react] No handler found for element ${ref.current?.id}`
+//         );
+//         return;
+//       }
+//       handler.setData(newData);
+//     },
+//     setMyAwareness: (newLocalAwareness) => {
+//       getCurrentElementHandler(
+//         TagType.CanPlay,
+//         ref.current?.id || ""
+//       )?.setMyAwareness(newLocalAwareness);
+//     },
+//     myAwareness,
+//     ref,
+//   };
+// }
+
+// @deprecated use withSharedState instead
 export const withPlay =
   <P extends object = {}>() =>
   <T extends object, V = any>(
@@ -137,35 +304,7 @@ export const withPlay =
     ) => React.ReactElement
   ) =>
     withPlayBase<P, T, V>(playConfig, component);
-
-// TODO: This is the ideal API but won't work until typescript supports
-// partial inference of generics. See https://github.com/microsoft/TypeScript/pull/26349
-// export function withPlay<
-//   P extends object = {},
-//   T extends object = {},
-//   V extends object = {}
-// >(
-//   {
-//     tagInfo,
-//     id,
-//     ...elementProps
-//   }: Omit<
-//     ReactElementInitializer<T, V> & {
-//       tagInfo?: Partial<{ [k in TagType]: string }>;
-//     },
-//     "children"
-//   >,
-//   component: (
-//     props: ReactElementEventHandlerData<T, V> & P
-//   ) => React.ReactElement
-// ): (props: P) => React.ReactElement {
-
-export type WithPlayProps<T, V> =
-  | Omit<ReactElementInitializer<T, V>, "children">
-  | (Omit<ReactElementInitializer<T, V>, "children" | "defaultData"> & {
-      tagInfo?: Partial<{ [k in TagType]: string }> | TagType[];
-    });
-
+// @deprecated use withSharedState instead
 export function withPlayBase<P, T extends object, V = any>(
   playConfig: WithPlayProps<T, V> | ((props: P) => WithPlayProps<T, V>),
   component: (
@@ -187,24 +326,6 @@ export function withPlayBase<P, T extends object, V = any>(
   // console.log("rendering", ref.current?.id, data, awareness, myAwareness);
   return (props) => renderChildren(props);
 }
-
-interface Props {}
-
-export const ToggleSquare = withPlay<Props>()(
-  { defaultData: { on: false } },
-  ({ data, setData, props }) => {
-    return (
-      <div
-        style={{
-          width: "200px",
-          height: "200px",
-          ...(data.on ? { background: "green" } : { background: "red" }),
-        }}
-        onClick={() => setData({ on: !data.on })}
-      />
-    );
-  }
-);
 
 export { playhtml };
 export { PlayProvider, PlayContext } from "./PlayProvider";